# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smaki CRM API Documentation",
        "version": "0.3"
    },
    "servers": [
        {
            "url": "https://api.smaki.com.ua/api"
        }
    ],
    "paths": {
...
        "/mobile/orders": {
            "get": {
                "tags": [
                    "Mobile.Orders"
                ],
                "summary": "Курьерское приложение. Получить список заказов для доставки",
                "description": "Доступ для пользователей с ролью: <b>Courier</b>",
                "operationId": "757135def13829fa121e85a6ec8281e3",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "restaurant": {
                                            "description": "Идентификатор ресторана",
                                            "type": "string"
                                        },
                                        "delivery_terminal_id": {
                                            "description": "Уникальный идентификатор доставочного терминала в iiko",
                                            "type": "string"
                                        },
                                        "status": {
                                            "description": "Статус заказа",
                                            "type": "string"
                                        },
                                        "order_uuid": {
                                            "description": "Уникальный идентификатор заказа",
                                            "type": "string"
                                        },
                                        "order_id": {
                                            "description": "Понятный номер заказа. Может использоваться для передачи клиенту. Уникальность не гарантирована",
                                            "type": "string"
                                        },
                                        "comment": {
                                            "description": "Комментарий к заказу",
                                            "type": "integer"
                                        },
                                        "expected_delivery_at": {
                                            "description": "Время, к которому нужно доставить заказ",
                                            "type": "integer"
                                        },
                                        "payment": {
                                            "properties": {
                                                "code": {
                                                    "description": "Тип оплаты",
                                                    "type": "string"
                                                },
                                                "title": {
                                                    "description": "Название типа оплаты",
                                                    "type": "string"
                                                },
                                                "sum": {
                                                    "description": "Сумма к оплате",
                                                    "type": "integer"
                                                },
                                                "prepareChangeFrom": {
                                                    "description": "Сдача с",
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "customer": {
                                            "properties": {
                                                "name": {
                                                    "description": "Имя клиента",
                                                    "type": "string"
                                                },
                                                "phone": {
                                                    "description": "Телефон клиента доставки",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "items": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "description": "Название позиции",
                                                        "type": "string"
                                                    },
                                                    "amount": {
                                                        "description": "Количество",
                                                        "type": "integer"
                                                    },
                                                    "sum": {
                                                        "description": "Сумма",
                                                        "type": "integer"
                                                    },
                                                    "comment": {
                                                        "description": "Комментарий",
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "address": {
                                            "properties": {
                                                "city": {
                                                    "description": "Наименование города",
                                                    "type": "string"
                                                },
                                                "street": {
                                                    "description": "Наименование улицы",
                                                    "type": "string"
                                                },
                                                "index": {
                                                    "description": "Индекс улицы в адресе, если есть",
                                                    "type": "string"
                                                },
                                                "home": {
                                                    "description": "Дом",
                                                    "type": "string"
                                                },
                                                "housing": {
                                                    "description": "Корпус",
                                                    "type": "string"
                                                },
                                                "apartment": {
                                                    "description": "Квартира",
                                                    "type": "string"
                                                },
                                                "entrance": {
                                                    "description": "Подъезд",
                                                    "type": "string"
                                                },
                                                "floor": {
                                                    "description": "Этаж",
                                                    "type": "string"
                                                },
                                                "comment": {
                                                    "description": "Дополнительная информация",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "data": [
                                            {
                                                "restaurant": "smaki",
                                                "delivery_terminal_id": "018cf4d4-0e07-b00a-0172-e4f18ac4ce92",
                                                "status": "waiting",
                                                "order_uuid": "257fd1c0-3015-dc6c-f751-d78dd06b4ef5",
                                                "order_id": 83112,
                                                "payment": {
                                                    "code": "CASH",
                                                    "title": "Наличные",
                                                    "sum": 609,
                                                    "prepareChangeFrom": 700
                                                },
                                                "comment": "Доставка за 29хв; | Замовлення онлайн | Підготувати решту з: готівка;",
                                                "expected_delivery_at": "2021-08-27 20:15:18",
                                                "customer": {
                                                    "name": "Юлія Кухар",
                                                    "phone": "+380974747560"
                                                },
                                                "items": [
                                                    {
                                                        "name": "Преміум Кілограм",
                                                        "amount": 1,
                                                        "sum": 609,
                                                        "comment": null
                                                    }
                                                ],
                                                "address": {
                                                    "city": "Львів",
                                                    "street": "Чернівецька вул.",
                                                    "index": "",
                                                    "home": "11",
                                                    "housing": "",
                                                    "apartment": "",
                                                    "entrance": "",
                                                    "floor": "2",
                                                    "comment": "Львів, Чернівецька , дім 11, поверх 2\\nЛьвів Чернівецька 11 2"
                                                }
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/mobile/orders/{order_uuid}": {
            "patch": {
                "tags": [
                    "Mobile.Orders"
                ],
                "summary": "Курьерское приложение. Изменить статус заказу из поездки",
                "description": "Отметить заказ доставленным или недоставленным. Доступ для пользователей с ролью: <b>Courier</b>",
                "operationId": "15126adb21cb72a1f88d446ba7a8c49e",
                "parameters": [
                    {
                        "name": "order_uuid",
                        "in": "path",
                        "description": "UUID заказа",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "restaurant": {
                                        "description": "Идентификатор ресторана",
                                        "type": "string"
                                    },
                                    "latitude": {
                                        "description": "Широта",
                                        "type": "string"
                                    },
                                    "longitude": {
                                        "description": "Долгота",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "success": {
                                                    "type": "boolean"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "security": [
        []
    ]
}